Feat: Course 도메인 예외처리 코드 작성

+++ src/main/java/coursemaker/coursemaker/domain/course/service/CourseServiceImpl.java
@@ -2,6 +2,8 @@ package coursemaker.coursemaker.domain.course.service;
 
 import coursemaker.coursemaker.domain.course.dto.*;
 import coursemaker.coursemaker.domain.course.entity.CourseDestination;
+import coursemaker.coursemaker.domain.course.exception.TravelCourseDuplicatedException;
+import coursemaker.coursemaker.domain.course.exception.TravelCourseNotFoundException;
 import coursemaker.coursemaker.domain.course.repository.CourseDestinationRepository;
 import coursemaker.coursemaker.domain.course.entity.TravelCourse;
 import coursemaker.coursemaker.domain.course.service.CourseDestinationService;
@@ -19,6 +21,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -35,6 +38,10 @@ public class CourseServiceImpl implements CourseService{
 
     @Override
     public TravelCourse save(AddTravelCourseRequest request) {
+        Optional<TravelCourse> existingCourse = travelCourseRepository.findByTitle(request.getTitle());
+        if (existingCourse.isPresent()) {
+            throw new TravelCourseDuplicatedException("이미 존재하는 코스입니다. ", "코스 이름: " + request.getTitle());
+        }
         return travelCourseRepository.save(request.toEntity());
     }
 
@@ -51,20 +58,31 @@ public class CourseServiceImpl implements CourseService{
     @Override
     public TravelCourse findById(long id) {
         return travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
+                .orElseThrow(() -> new TravelCourseNotFoundException("존재하지 않는 코스입니다. ", "Course ID: " + id));
     }
 
     @Override
     public TravelCourse update(long id, UpdateTravelCourseRequest request) {
         TravelCourse travelCourse = travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
+                .orElseThrow(() -> new TravelCourseNotFoundException("수정할 코스가 존재하지 않습니다. ", "Course ID: " + id));
         travelCourse.update(request.getTitle(), request.getContent(), request.getDuration(), request.getTravelerCount(), request.getTravelType(), request.getPictureLink());
         return travelCourseRepository.save(travelCourse);
     }
 
     @Override
     public void delete(long id) {
+        if (!travelCourseRepository.existsById(id)) {
+            throw new TravelCourseNotFoundException("삭제할 코스가 존재하지 않습니다. " + id, "Course ID: " + id);
+        }
         travelCourseRepository.deleteById(id);
+    }
+
+    @Override
+    public TravelCourse incrementViews(long id) {
+        TravelCourse travelCourse = travelCourseRepository.findById(id)
+                .orElseThrow(() -> new TravelCourseNotFoundException("코스가 존재하지 않습니다.: ", "Course ID: " + id));
+        travelCourse.incrementViews();
+        return travelCourseRepository.save(travelCourse);
     }
 
 //    @Override
@@ -99,12 +117,4 @@ public class CourseServiceImpl implements CourseService{
 //    public void deleteCourseDestination(long id) {
 //        courseDestinationRepository.deleteById(id);
 //    }
-
-    @Override
-    public TravelCourse incrementViews(long id) {
-        TravelCourse travelCourse = travelCourseRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Travel course not found with id: " + id));
-        travelCourse.incrementViews();
-        return travelCourseRepository.save(travelCourse);
-    }
 }
